requirements.txt
discord.py==2.1.*
openai==0.26.*